ingested
